+++ .feature-model
@@ -1,4 +1,5 @@
 HAnSViz
+    JSONHandler
     HAnSData
         Service
     Plotting
