+++ gradle.properties
@@ -1,7 +1,7 @@
 # IntelliJ Platform Artifacts Repositories -> https://plugins.jetbrains.com/docs/intellij/intellij-artifacts.html
 
-pluginGroup = com.github.johmara.hansviz
-pluginName = HAnS-viz
+pluginGroup = se.isselab.hansviz
+pluginName = HAnS-Viz
 pluginRepositoryUrl = https://github.com/isselab/HAnS-viz
 # SemVer format -> https://semver.org
 pluginVersion = 0.0.1
