--- src/main/java/JSONHandler/JSONType.java
@@ -1,8 +1,0 @@
-package JSONHandler;
-
-public enum JSONType {
-    Default,
-    Tree,
-    Treemap,
-    Tangling
-}
