--- src/main/java/com/github/johmara/hansviz/browser/.feature-to-file
@@ -1,2 +1,0 @@
-BrowserResourceHandler.java BrowserSchemeHandlerFactory.java BrowserViewerService.java BrowserViewerWindow.java BrowserWindowFactory.java ClosedConnection.java OpenedConnection.java ResourceHandlerState.java
-Browser
