--- src/main/java/com/github/johmara/hansviz/browser/jshandler/.feature-to-file
@@ -1,2 +1,0 @@
-JSMessageRouterHandler.java
-JSHandler
