+++ src/main/java/se/isselab/hansviz/JSONHandler/.feature-to-folder
@@ -1,0 +1,1 @@
+JSONHandler
