+++ src/main/java/se/isselab/hansviz/JSONHandler/JSONType.java
@@ -1,0 +1,8 @@
+package se.isselab.hansviz.JSONHandler;
+
+public enum JSONType {
+    Default,
+    Tree,
+    Treemap,
+    Tangling
+}
