--- src/main/kotlin/com/github/johmara/hansviz/MyBundle.kt
@@ -1,20 +1,0 @@
-package com.github.johmara.hansviz
-
-import com.intellij.DynamicBundle
-import org.jetbrains.annotations.NonNls
-import org.jetbrains.annotations.PropertyKey
-
-@NonNls
-private const val BUNDLE = "messages.MyBundle"
-
-object MyBundle : DynamicBundle(BUNDLE) {
-
-    @JvmStatic
-    fun message(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any) =
-        getMessage(key, *params)
-
-    @Suppress("unused")
-    @JvmStatic
-    fun messagePointer(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any) =
-        getLazyMessage(key, *params)
-}
