--- src/main/kotlin/com/github/johmara/hansviz/listeners/MyApplicationActivationListener.kt
@@ -1,12 +1,0 @@
-package com.github.johmara.hansviz.listeners
-
-import com.intellij.openapi.application.ApplicationActivationListener
-import com.intellij.openapi.diagnostic.thisLogger
-import com.intellij.openapi.wm.IdeFrame
-
-internal class MyApplicationActivationListener : ApplicationActivationListener {
-
-    override fun applicationActivated(ideFrame: IdeFrame) {
-        thisLogger().warn("Don't forget to remove all non-needed sample code files with their corresponding registration entries in `plugin.xml`.")
-    }
-}
