--- src/main/kotlin/com/github/johmara/hansviz/services/MyProjectService.kt
@@ -1,17 +1,0 @@
-package com.github.johmara.hansviz.services
-
-import com.intellij.openapi.components.Service
-import com.intellij.openapi.diagnostic.thisLogger
-import com.intellij.openapi.project.Project
-import com.github.johmara.hansviz.MyBundle
-
-@Service(Service.Level.PROJECT)
-class MyProjectService(project: Project) {
-
-    init {
-        thisLogger().info(MyBundle.message("projectService", project.name))
-        thisLogger().warn("Don't forget to remove all non-needed sample code files with their corresponding registration entries in `plugin.xml`.")
-    }
-
-    fun getRandomNumber() = (1..100).random()
-}
