--- src/test/testData/rename/foo.xml
@@ -1,3 +1,0 @@
-<root>
-    <a<caret>1>Foo</a1>
-</root>
