--- src/test/testData/rename/foo_after.xml
@@ -1,3 +1,0 @@
-<root>
-    <a2>Foo</a2>
-</root>
